feat: migrate money to Long cents with tested business rules

+++ app/src/main/java/com/confeitaria/gestao/data/local/database/dao/PedidoDao.kt
@@ -28,5 +28,5 @@ interface PedidoDao {
     fun countAtivos(): Flow<Int>
 
     @Query("SELECT SUM(totalFinal) FROM pedido WHERE status = 'ENTREGUE' AND dataEntrega BETWEEN :inicio AND :fim")
-    fun getReceitaPeriodo(inicio: Long, fim: Long): Flow<Double?>
+    fun getReceitaPeriodo(inicio: Long, fim: Long): Flow<Long?>
 }
